requirements.txt
appdirs==1.4.0
decorator==4.0.11
elasticsearch==1.9.0
packaging==16.8
pyparsing==2.1.10
python-dateutil==2.6.0
PyYAML==3.12
requests==2.13.0
-e git+https://github.com/EHRI/resync@ehribranch#egg=resyncehri
-e git+https://github.com/EHRI/rspub-core@master#egg=rspub-core
six==1.10.0
urllib3==1.20
validators==0.11.2
